backend/requirements.txt
django
djangorestframework 
djangorestframework-simplejwt 
django-cors-headers
six
Pillow
psycopg2
psycopg2-binary